+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Enh: allow to get engine version synchronously.
 - Enh: APIS-592, Allow to pass an object to createCUBRIDConnection().
 - Enh: allow to pass a connection timeout value as part of connection parameters.
+- Enh: no need to quote a number.
 - Fix: test path for `node_modules`.
 - Fix: host name for tests.
 - Fix: travis script.